+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 * feat: class attributes are now stored in the metadata
 * chore: rector run
 * feat: collect attributes if using property path mapping
+* fix: mapping to property without setter wasn't working
 
 ## 1.8.0